Modified how the setSamplingInterval function works.
As well as the getTimebase function.

The getTimebase now returns two things, the raw timebase number required
by the picoscope, and a human readible time interval.

This used by the setSampling interval function to compute the total
number of samples needed. It returns this to the user.

Gah, the setSigGenBuiltIn function is messed up, and doesn't really
work. I will have to followup on the bug report I posted earlier.