Wow. spent all day debugging the AWG
It turns out that their programming guide is wrong.

They claimed that -32768 was the lowest possible value in
the AWG and that  +32767 was the largest.

It turns out that 0x0000 is the lowest possible value and that
                  0x0FFF is the largest

This commit doesn't fix the code but rather shows a demo of this
working.

I'll fix the code in the next update