So I ran the stress test.
In the test, a new array is allocated every loop, for 4000 iterations and we take 4096 data points.

It seems to work very well (and fast may I say....)

I compiled this by overwriting their PS6000con file.

So there is only one other difference. Possibly the driver version could be different than the one
that is loaded via python.

The sdk version is PS6000sdk_r10_5_0_28

If only I knew how to chose which DLL was included.

I'll have to spend more time debugging this.