AWGMin/MaxVal correction
AWG scaling according to programming manual p.72:
Vout = 1uV * (pkToPk/2) * (sample_value / 32767) + offsetVoltage
Furthermore, the API datatype is a (signed) short. I could not get negative voltages until I changed the minimum value accordingly to -32767

The original code refers to an issue for the PS6000, where a value of 4096 seemed appropriate. The change was tested on my 2206A so I suppose it was specific for that scope?