getRawDataBulk: clear the driver's data pointers after use
Fixes #149.

The driver's GetValuesBulk routine will retrieve data on all channels
for which a pointer has been set with SetDataBuffer. This means that
calling getDataRawBulk on two different channels will result in the
second call writing to the buffer from the first, which may no longer
been valid.

The fix is to zero these in the driver before returning, just as
getDataRaw() does.