Checked for completeness, methods improved.
All methods of the manual are present (at least as stubs). Methods copied from the ps5000 are marked as not present.
Some methods, values are improved, according to the manual. Order improved.
Some new methods (updates etc.) included.

A test-file was included in order to test the methods with the picoscope.